Move PHPUnit tests to tests/unit, to distinguish between upcoming integration tests
Later, there should be integration tests,
so prepare a separate space for them.
