+++ tests/php/unit/admin/test-class-admin.php
