+++ tests/php/unit/admin/test-class-license.php
