+++ tests/php/unit/admin/test-class-settings.php
