+++ tests/php/unit/admin/test-class-upgrade.php
