+++ tests/php/unit/blocks/controls/test-class-checkbox.php
