+++ tests/php/unit/blocks/controls/test-class-color.php
