+++ tests/php/unit/blocks/controls/test-class-control-abstract.php
