+++ tests/php/unit/blocks/controls/test-class-control-setting.php
