+++ tests/php/unit/blocks/controls/test-class-image.php
