+++ tests/php/unit/blocks/controls/test-class-number.php
