+++ tests/php/unit/blocks/controls/test-class-repeater.php
