+++ tests/php/unit/blocks/controls/test-class-rich-text.php
