+++ tests/php/unit/blocks/controls/test-class-text.php
