+++ tests/php/unit/blocks/controls/test-class-toggle.php
