+++ tests/php/unit/blocks/controls/test-class-user.php
