+++ tests/php/unit/blocks/test-class-block.php
