+++ tests/php/unit/blocks/test-class-field.php
