+++ tests/php/unit/blocks/test-class-loader.php
