+++ tests/php/unit/blocks/test-class-loop.php
