+++ tests/php/unit/helpers/class-abstract-template.php
