+++ tests/php/unit/post-types/test-class-block-post.php
